convert datetime to str if driver converts silently anyway

+++ fiona/rfc3339.py
@@ -54,7 +54,7 @@ def parse_time(text):
             int(g.group(1)),
             int(g.group(3)),
             int(g.group(5)),
-            1000000.0 * float(g.group(6)))
+            int(1000000.0 * float(g.group(6))))
 
 
 def parse_date(text):
@@ -85,4 +85,4 @@ def parse_datetime(text):
         int(g.group(7)), 
         int(g.group(9)), 
         int(g.group(11)),
-        1000000.0*float(g.group(12)))
+        int(1000000.0*float(g.group(12))))
